# api-catalogo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servicio de Catálogo de Homelab Collector Platform.",
        "title": "Homelab Collector Platform - Catalogo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9094",
    "basePath": "/",
    "paths": {
...
        "/api/v1/catalogo/estados-arbol": {
            "get": {
                "description": "Obtiene la lista de todos los estados de árbol registrados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estados de Árbol"
                ],
                "summary": "Listar estados de árbol",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.EstadoArbol"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un nuevo estado de salud/situación del árbol en el catálogo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estados de Árbol"
                ],
                "summary": "Crear estado de árbol",
                "parameters": [
                    {
                        "description": "Datos para crear el estado de árbol",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateEstadoArbolRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.EstadoArbol"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/catalogo/estados-arbol/{id}": {
            "get": {
                "description": "Obtiene los detalles de un estado de árbol por su UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estados de Árbol"
                ],
                "summary": "Obtener estado de árbol por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del estado (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.EstadoArbol"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza los datos de un estado de árbol existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estados de Árbol"
                ],
                "summary": "Actualizar estado de árbol",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del estado (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos a actualizar",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateEstadoArbolRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.EstadoArbol"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un estado de árbol del catálogo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estados de Árbol"
                ],
                "summary": "Eliminar estado de árbol",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del estado (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateEstadoArbolRequest": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
...
        "model.EstadoArbol": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateEstadoArbolRequest": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
...
        "response.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.APIError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}